+++ cpp/include/cuml/ensemble/randomforest.hpp
@@ -146,8 +146,9 @@ void compare_concat_forest_to_subforests(
   ModelHandle concat_tree_handle, std::vector<ModelHandle> treelite_handles);
 
 template <class T, class L>
-std::vector<double> obtain_forest_info(const RandomForestMetaData<T, L>* forest,
-                                       int param_requested);
+std::vector<std::vector<double>> obtain_forest_info(
+  const RandomForestMetaData<T, L>* forest,
+  int param_requested);
 
 template <class T, class L>
 int calc_num_nodes(const RandomForestMetaData<T, L>* forest);