+++ python/cuml/ensemble/randomforestregressor.pyx
@@ -432,7 +432,7 @@ class RandomForestRegressor(Base, RegressorMixin):
         self.treelite_serialized_model = treelite_serialize(model_ptr)
         return self.treelite_serialized_model
 
-    def total_nodes_(self):
+    def n_nodes_(self):
         """
         Calculates the number of nodes in the trained RF model
 
@@ -474,17 +474,20 @@ class RandomForestRegressor(Base, RegressorMixin):
         select_feature = {"threshold" : 1,
                           "best_metric" : 2,
                           "column_id" : 3}
-        cdef RandomForestMetaData[float, int] *rf_forest = \
-            <RandomForestMetaData[float, int]*><uintptr_t> self.rf_forest
-        cdef RandomForestMetaData[double, int] *rf_forest64 = \
-            <RandomForestMetaData[double, int]*><uintptr_t> self.rf_forest64
+        cdef RandomForestMetaData[float, float] *rf_forest = \
+            <RandomForestMetaData[float, float]*><uintptr_t> self.rf_forest
+        cdef RandomForestMetaData[double, double] *rf_forest64 = \
+            <RandomForestMetaData[double, double]*><uintptr_t> self.rf_forest64
+        print("select_feature[str(forest_param)] :  ", select_feature[str(forest_param)])
 
         if self.dtype == np.float32:
-          forest_info_vector = obtain_forest_info(rf_forest,
-                                                  select_feature[str(forest_param)])
+          forest_info_vector = obtain_forest_info(
+            rf_forest,
+            <int> select_feature[str(forest_param)])
         else:
-          forest_info_vector = obtain_forest_info(rf_forest64,
-                                                  select_feature[str(forest_param)])
+          forest_info_vector = obtain_forest_info(
+            rf_forest64,
+            <int> select_feature[str(forest_param)])
 
         return forest_info_vector